Initial implementation of write testing
This implements basic write testing with unhandled write detection. This
is imperfect since it may sometimes report wrong location: if there's a
correct write following incorrect one that happens to write same data
the latter one will be reported.

As an example `write(&[1, 0])`, `write(&[0])` will trigger this.
To address the issue a search algorithm must be used similar to what
read does. I initially thought it doesn't need to be used and only
realized it when it was almost finished. Partial implementation is still
much more useful than no implementation at all.