(Ab?)use `std::panic::Location` for error location
Since Rust 1.46 we can use `std::panic::Location` to get the location of
the caller as long as we mark functions with `#[track_caller]`. This
saves us from compiling `backtrace` crate which takes some time and from
manually searching backtrace.

The change supports this behind a feature to help crates with older
MSRV. If both features are enabled `backtrace` is used because it's
richer.